Update patch set 2
Patch Set 2:

So I was testing this and it failed for me just now, then I realised that I was testing with a first revision (which with this having its current base wont work)
I moved that to the next revision and everything worked as expected.
I'm guessing after a rebase your issue will be solved!

Patch-set: 2
Label: Code-Review=0