Update patch set 2
Patch Set 2:

I kind of feel like this should work slightly differently.
The switching of dragging the bottom pointer to the top is now super odd.

I think when switching the pointer you are dragging should remain on the same side of the slider (in regards to top and bottom)

Patch-set: 2
Label: Code-Review=0