Update patch set 2
Patch Set 2:

I agree.
But then, if I didnt get intention behind adding the shadow wrong and now I am not failing in expressing it: the point in adding the shadow was to sort of distinguish between "inactive" elements like the bars, and "active" (draggable) pointers, and to make it clear to the user that those tiny things with a shadow are the thing to move around.

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0