Update patch set 3
Patch Set 3:

There's an inconsistency between what moment displays (local time for the user) and the output that's expected in the tests (exact timestamp, UTC).

Other than asking the tests the result through moment (which seems to be redundant) I'm not entirely sure what the best way to fix this is right now, I'm going to consult the moment documentation. In any case, using moment should be a lot better than manually constructing the timestamp - for i18n reasons and for consistency with mediawiki development in general (moment is a core library, and is used widely)

One option would be to "trick" moment -- in the tests only -- to consider the user's timezone as if it is UTC. That would make the hard-coded test timestamp strings work.

Patch-set: 3
Label: Verified=0