Update patch set 4
Patch Set 4: Code-Review-1

Nice thing with momentjs but the patch removes the offset calculation we need to have the same times for the tooltip and the diffpage. 

You can test that by changing your timezone in the user settings and comparing the times of the revisions on the diff page and the ones shown in the toopltip.

Patch-set: 4
Reviewer: WMDE-Fisch <1883@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1