Update patch set 4
Patch Set 4:

Hm, I was sure I tested it, but I may have missed soemthing.

Moment is supposed to be able to deal with this - we've been using it in Echo, too, and there it's using the user's local time by default, rather than UTC. Now that I am thinking about this, though, we may have been asking the API for mw timestamp and not UTC, so we're already getting it localized.

There is another option: We could keep the offset option, and give the formatDate method a zone() calculation:

return moment( rawDate ).zone( offset ).format( 'HH:mm, D MMM YYYY' );

That should fix things if we are expecting the timestamp to be UTC.

The only thing that I am a little meh about in this is that the tests are changing the user option directly (I personally prefer to avoid impacting global variables in tests) -- but we can solve this also by creating a class variable "offset" in RevisionSlider.Revision that gets the user offset on init but can be manipulated and changed in tests.

I'll work on this and submit another patch.

Patch-set: 4
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0