Update patch set 3
Patch Set 3: Code-Review-1

This appears to break more things than it fixes for me!

The most recent revision still does not appear to be visible, but instead the pointers point to the revision prior to the one that they probably should.

As a result it is also impossible to make either pointer point to either of the 2 most recent revisions.

Patch-set: 3
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1