Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

I'm not sure why the tests are padding but this is still not working for me locally..

So while my server time is 11:49 and my local time is set as 14:49 so a +3 difference.

While viewing as anon and using the slider a time in the diff appears as 2:26 and in the tooltip as 2:26.

As a logged in user with a timezone of -4 I see 19:26 in the diff but 2:26 in the tooltip still.

So when logged in the timezone is being adjusted for the server offset / default offset but not my own user offset?

Patch-set: 7
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1