Update patch set 2
Patch Set 2:

(1 comment)

I like dropping the "stopper" divs! Slider (pointer positions etc) still look fine for me. It would be nice if others could also have a look.

Having tested it briefly it looks like "pagination" is not fully working in RTL mode. Arrows work (and pointers also seem to be affected correctly) but the bar plot does not change after going forwards or backwards. For LTR it works OK, so there seems to be some bit missing in the RTL-specific code. I might try to track it down.

Also one possibly completely wrong remark: Unfortunately I don't speak Hebrew or any other language that is written right-to-left. But isn't it that punctuation also goes right-to-left as rest of the text? At least colons in diff headers are place left-most. But then "Revision Slider is loading" (same for noscript message) has a period on the right-most end of the sentence. Are those messages displayed properly (I cannot read what message says so I am just guessing based on the position of the period).
It might also just be because my wiki is generally LTR but I only "force" RTL by setting uselang=he in URL.
And sorry if all is correct and all the stuff above is completely foobar :)

Patch-set: 2
Label: Verified=0