Update patch set 4
Patch Set 4: -Verified Code-Review-1

Changing +1 to -1 as I found a regression!
This patch should also update Slider's isAtEnd method. Now it shows the "arrow forwards" even when I am at the last "pagination window". In the condition in the method mentioned there should be
this.getLastVisibleRevisionIndex() === this.revisions.getLength() - 1
instead of 
this.getLastVisibleRevisionIndex() === this.revisions.getLength() - 2

Patch-set: 4
Label: Code-Review=-1