Update patch set 3
Patch Set 3:

Yeah I thought of abstracting the rtl check, but it didn't seem urgent because basically it would check a static property... either $( 'body' ).css( 'direction' ) or some code variable we're caching this value into, and there's not a lot of difference. 

Usually, the way we check directionality also differentiates between user's interface and wiki's content direction (both of which can be different to one another) - but in this case, I'm not sure we need to worry about content at all, which is another reason why I didn't think caching the value is crucial.

Then again, it could be good for consistency. I have no preference, if you guys prefer it we can definitely do that. It can also help with the tests of 'is rtl' rather than 'is ltr' etc.

I'll try to go over this to make sure at least the base cases are covered in RTL after the rebase and +1 if I think it's good. I think it's okay if the behavior is not 100% in RTL, though -- this can be a first step. If we continue tweaking/waiting then we'll have to continuously rebase it, and that will be tedious. There are probably things we can fix in (smaller) followup commits.

Patch-set: 3
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0