Update patch set 3
Patch Set 3:

Hi Nemo_bis

it's probably hard to find something gender-neutral that fits for every language, but we are optimistic that translators find individual alternative solutions that suite the task.

For example: It might not work using the verb in Russian but for what my dictionary tells me, there's a gender-neutral form of the Russian word for "Username", for "User" and for "Name", so translators could use one of these and they have the freedom to do so.

Patch-set: 3