Update patch set 3
Patch Set 3:

I am not sure whether the discussion is supposed to happen here or in T136367 (there are some voices raised in both places). I'll try here.

I am not sure what "key" suggests better what the message is about, "Username" or "Edited-By". Nevertheless what might be an improvement is rephrasing the description in qqq.json a bit, so it didi not explicitly require "Gender-neutral label" but stated that usage of "GENDER is NOT supported" instead.
I believe all agree translators are smart and will figure out what to use in case of their language. The point is that current description qqq.json tends to require the gender-neutral label. There are languages where it might not possible which leads to use gender-specific variants. As we are not able to use GENDER for performance reasons, I'd consider mentioning this explicitly and letting translators decide what words would they choose.

BTW, WMDE-Fisch, I am not sure what Russian translation of "username" you've looked up but I believe there might be asimilar situation in Russian as you mentioned from Spanish, ie. there are different words for "name of the male user" and "name of the female user" in Russian as well.

Patch-set: 3
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0