Update patch set 1
Patch Set 1:

(2 comments)

Had a brief look and can confirm that it works (I have not looked at the content of the dialogue etc).
I've left some comments inline. Also what could be considered is cleaning Inkscape SVGs. I don't know what is the general MediaWiki policy for this, though. For instance, I think there is not need for having <sodipodi:namedview> elements here. Or is it?

Patch-set: 1
Label: Code-Review=0