Update patch set 2
Patch Set 2:

I haven't looked at the code too closely but the approach seems fine to me and it works quite well already.

I think it would be helpful to show a loading indicator either inside the arrow or next to the revision that was last loaded. Also, another way to deal with the jumping revision size would be to show it only once the previous revision was loaded although that might look strange as well. Both of these points should probably be addressed in separate patches though.

Patch-set: 2
Label: Code-Review=0