Allow browser tests to add more revisions easily
This allows an arbitrary number of revisions to be created
on a page very easily.
The page created will have a name relating to RevisionSlider
suffixed by a random float.

The base logic for this is taken from:
Ic165f108517c534a4b1c12883621b3624c7b0336

This patch also removes the @vagrant tags as RevisonSlider
is not yet in vragrant.
Please see T139296 for the ticket to add it.

Bug: T133278
Change-Id: I8a25db411bee2fa44555a23d1a840907a185c998