Update patch set 3
Patch Set 3: Verified+1 Code-Review+1

(1 comment)

Awesome, works really well!

I’m not too excited about the code duplication for getting and adding newer or older revisions in SliderView.js but I see that the behavior is quite different so I guess unless we come up with something good it’s better left as a duplication than a bad spaghetti abstraction. I’m also getting a bit scared about the size and amount of things done in SliderView.js in general but that’s also not something that should be discussed in too much detail here. It would probably be nice to factor the code for the added functionality out of SliderView.js but I’m not entirely convinced it’ll be worth it.

To sum it up I think this patch is well done and works well so I wouldn’t be against merging it as is.

Patch-set: 3
Reviewer: Jakob <1786@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=+1
Label: Code-Review=+1