Update patch set 3
Patch Set 3:

Adam, I think what you're describing is what Leszek pointed out in the commit message for patch set 1. The height of the revision bar depends on how many bytes were added after the previous revision. Since the previous revision is not fetched at that point the bar is quite high because it takes the revision's size as the relative size. It then jumps once the previous revision gets loaded and the actual relative size gets calculated.

Patch-set: 3