Update patch set 3
Patch Set 3:

Addshore, Jakob: yes, I've described this problem in the commit message of PS1 (along with some other todos some of which I've managed to solve) as it looked not that nice to have such a long commit message. Sorry for that, I should have made concerns I have more prominent.

But to the point: now that I am thinking about, with loading "older" revisiosn to avoid the jummping size of the "current oldest" review bar, what could to some extent help would be fetching one older revision more. It would not get added to the revisions chart but would give current bar size information for the oldest revision actually added. I hope it is possible to understand this long sentence.
This would make the whole stuff even more nasty. though.

Patch-set: 3