Update patch set 6
Patch Set 6:

Yes, this is a known problem (it used to be listed as a TODO in the commit message of the first patch set).

I think this is difficult to have a general solution for this, although I might have an idea for a trick to handle such cases (dynamically adjusting "revision-container's'" width. This will still be a bit ugly. But let's try this out.

Patch-set: 6
Label: Code-Review=0