Update patch set 1
Patch Set 1: Code-Review-1

When I discussed the issue with Jan he strongly suggested to have ALL the tool-tips, either above OR underneath the bars. 

The reason for that is, that the user probably kind of scrolls through all the revisions / tool-tips looking for something interesting and when the tool-tips jump from top to bottom it gets quite irritating. 

So the idea is to check the available space once when loading and then set a top / bottom mode.

Patch-set: 1
Reviewer: WMDE-Fisch <1883@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1