Update patch set 2
Patch Set 2:

So the selenium tests failure seems to have been some CI hiccup.

And per Jakob's comment: this thing is called an "arrow" in tipsy's code :) Which kind of conflicts with RevisionSlider's vocabulary

But to the point: I've seen this issue happening before. It seems to me it is a bug/issue in tipsy. It always tries to place this "arrow" in a center of the tooltip box but in a way so it is also vertically aligned just over/under/whatever the related element. That makes sense for most cases but when tooltip is too big or pushed to edge of the window this might just not possible to position the arrow correctly and it gets hidden/covered. At least this is the way I got it when I was trying to track down what's exactly happening there.
It seems to me this should rather be addressed in tipsy, other solutions would be some rather nasty JS hacks. I am really not willing to go this path

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0