Update patch set 1
Patch Set 1:

> I hope translators are not totally annoyed by our constant changes
 > ;-)

This is OK. It's understandable because this is a new feature, you are not changing a lot of messages, you are making them easier to translate, and you have good qqq documentation.

It is annoying when it happens to many dozens of messages repeatedly, as it already happened in some extensions in the past, but it's really not the case here.

Patch-set: 1