Update patch set 10
Patch Set 10: Code-Review-1

Demonstrated in a hangout now.
It looks like when clicking on the revisions in some cases the pointers do not swap when they should.
This results in the blue pointer (which should always be on the right) ending up on the left.
Thus the CSS for the coloured bars does not work.

Patch-set: 10
Label: Code-Review=-1