Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Tiny bit in browser tests must be fixed. Looks good apart from that.

I like getting rid of "darkness". What kind of darkness is it when it is white?

I am not sure what you mean by "The jQuery for appending the 'darkness' mask was broken" in the commit message. If including a <div> inside a table row this was indeed not very elegant and AFAIR it was only added to have it look OK in IE (adding <div> directly to <table> was not working there). With change you propose it should no longer be an issue, I guess.

Patch-set: 2
Label: Code-Review=-1