Update patch set 2
Patch Set 2:

Taking back my previous comment (partly).

So I like the change of name and simplifying the style.
But we still need to have a <div> inserted over the table.diff[data-mw="interface"]
This is done so when user click something in the slider, and the diff is being reloaded, during this user should not be able to click on things in the diff table, e.g. "← Older edit" as this would reload the whole page

Patch-set: 2