Also track the hook run count
This will give us a better idea of how many times the
slider is being sent out in by the php code as oposed
to how many times the JS code is running.

This number is important as client side the code will
listed to things like DoNotTrack and not send data!

The difference between the initphp and init numbers
should give us an idea of how often this is heppening!

Change-Id: I01db4318b08851de4c6b8c96cbf2740324160140