Update patch set 3
Patch Set 3:

Basically if user was allowed to click on another revision bar while diff has not yet reloaded we wouldn't be able which of API call for a new diff is the "right" one, and which should be ignored.

If the API call fired later responded before the one that was fired earlier wrong diff would be shown.

Patch-set: 3
Label: Verified=0