Update patch set 3
Patch Set 3:

Thanks to Tobias Gritschacher I realized that browser tests fail because dependency on OOjs UI patch is not working (PHP ToggleButtonWidgets are not found and pages simply crash).

It seens that dependency on OOjs UI change cannot be done in such a simple way. I believe the OOJS UI patch(es) must first be merged, then the new release of the library containing those changes must be released, and finally a core patch updating the OOJS dependency will be created.
This patch can only depend on the last element of those listed above, so browser tests for this will be failing on Jenkins as long as the OOJS changes are not merged (and released).

Unless I am simply mistaken and there is a simple way to have them passing now.

Browser test might be already run locally but this requires manual checking out and installing of OOJS UI patch.

Patch-set: 3
Label: Verified=0