Update patch set 1
Patch Set 1: Code-Review-1

I think ext.RevisionSlider.SliderView line 46 would now need a HelpDialog.init() before HelpDialog.show() in order to reset the slideshow (probably other places too?).

Because right now if you reopen the slideshow it would start with the last slide that was visible when it was closed. So if you open the slideshow, click next and close it (and do this multiple times in a row) it would crash after the 4th time.

Patch-set: 1
Label: Code-Review=-1