Update patch set 2
Patch Set 2:

Looks good.
Would you also like to switch to mw.log.log/warn/idontknow in DiffPage.js wher it adds a hint related to WikEdDiff?
This patch could be merged as it is now, but if you wanted to add the other change to it too, please comment fast before I merge it! :)

Patch-set: 2
Label: Verified=0