Update patch set 13
Patch Set 13: Code-Review-1

Well done. In most cases everything seems to work fine now, but sometimes when I drag and abort the pointer is moving back to a position a bit above or below the grid o.O

Looks a like a fault by jqueryui draggable when calculating the offsets. The only solution I came up with at the moment would be re-rendering the whole SliderView on escapePressed by calling `self.render( $container );`... but I am not sure if that is a bit too much ....

Patch-set: 13
Label: Code-Review=-1