Update patch set 13
Patch Set 13:

Hi Divadsn, thanks for also looking into this.

A posted a video in the Phabricator ticket here:

https://phabricator.wikimedia.org/T140965#2944610

It seems that jqueryui.draggable calculates wrong positions when dragging and aborting multiple times. - Sadly updating to a newer version of jqueryui.draggable is not trivial.

I also tried to reorder setting the pointer positions, the only thing that really helped was re-rendering the whole slider view. 

One thing that might work would be implementing a new method for correctly repositioning the pointers ( X and Y Axis ).

Patch-set: 13