Update patch set 2
Patch Set 2:

So if I managed to track the reason why this 4px of extra space around pointers has been introduced was to make it easier for users to drag'n'drop pointers (T143091). I am not good with such intricate fiddling, so I might not be accurate but this change does not seem to be making dragging pointers less easy. I haven't tested it too "deep", though.

I am not sure if this is the same issue that causes browser tests to fail but when I check this patch locally, the "box" around the pointer element seems to be a bit shifted. Namely, it looks few pixels off towards right (in a LTR UI). Because of that I cannot click the the revision bar right to the revision pointers points at, I mean the left part of this bar, it is kind of covered by the pointer's "box". The right part of the rev bar is clickable. So there is some issue with margin etc calculation apparently.

Patch-set: 2