Remove browser tests for beta feature setting
The browser tests do not gain very much from the beta
feature tests and the default setting having the extension
only availible as a beta feature slows down the tests in
general.

Furthermore removing this will make it easier to let tests
pass for the temporary change in the slider logic.

Change-Id: I7377a6f11aa103af76d14fea3cbb34572495192a