Update patch set 1
Patch Set 1: Verified-1

Hmm when I try this locally I get different results in the tooltips on master vs this patch. On master the tooltip time fits the time shown on the diff page. With this patch I have an offset of 2h. ( what seems to be the UTC offset setting I have in the user preferences ). 

So either it did not work correctly on master or it does not work correctly with this patch. Others should check with their settings. It probably also depends on the setting of the time zone the server is set to.

Patch-set: 1
Label: Verified=-1