Do not use the main config for extension configuration options
This currently just works, because the currently used backend are
global variables. Once the main config uses a more scoped/focused
backend, such as a HashConfig or a database-based config, requesting
extension configurations from it will not work anymore.

Therefore, use the extensions Config object to request configuration
options for the extension.

Change-Id: Iea02d3c284f9a0e12abff33e3624dce9585ab51c