Fix errors leading to wrong slider scroll postions
The detection touched in this patch addresses slightly different
kinds of RTL handling in browsers. The detection how things are
handled seems to be broken in the mentioned Chrome V8 versions.

The workaround takes care of that issue. With Chrome V8 6.0 this
should be fixed and the workaround can be removed.

P.S.: Credits to Leszek for the idea what's actually going on.

Bug: T168299
Change-Id: I068c2fb6ba60189004c0ec6661fcd346af16126c