Update patch set 1
Patch Set 1: Code-Review-1

Hey thanks for the work here ... due to the slight changes in the colors with the auto computation I checked back with UX/design.

The yellow for the pointer knobs is really too bright now. When they originally fiddled with the color spectrum they also ran into that problem and changed it to be a bit more reddish. So we might re-add a third color here again.

Blue is fine. :-)

Patch-set: 1
Label: Code-Review=-1