Update patch set 3
Patch Set 3: Code-Review-2

Peter, sorry for the discussion here. I think we as product owners have to talk about the general issue a bit more to come to a solution we are all happy with.

I think the patch goes into the right direction considering what is requested in the ticket. Going with with a fixed duration again will not solve the issue described there. If what requested in the ticket is not what we want to have that is another issue.

We should probably include UX to find a good solution and then can get back to this patch later.

Thanks for the help so far.

Patch-set: 3