Update patch set 3
Patch Set 3:

(3 comments)

Sooo after a longer consultation with PM and Math Experts we came to several conclusion that should work making the animation great again!

Sorry for all the hassle but we were not aware of the pitfalls with that task and how tricky it can be to find the function that feels and looks right calculating the actual animation time.

Patch-set: 3